make extract() match two interfaces
while making layer pagination work I found the result of the pagination must implement the ResultSetInterface. A simple stub trait did this for Stack Sets. But Stack entities are not natural result sets. They had an implentaion of extract() in the entity interface that is incomplatible with the result set interface.

So I hacked the signature of the stub method to make it match both interfaces.